recommendation-ai/requirements.txt
# ADK with A2A and extensions (LiteLLM) extras
google-adk[a2a,extensions]>=1.0.0

# Session store
redis>=5.0.0

# Environment variable loading
python-dotenv>=1.0.0

# ASGI server
uvicorn>=0.29.0

# HTTP client (used internally by MCP toolset)
httpx>=0.27.0

# Data validation
pydantic>=2.7.0

# MCP server SDK (for the local dev stubs)
mcp>=1.0.0